requirements.txt
discord.py==1.7.3
google-api-python-client==2.34.0
